feat: IFS-2488: move isy-security related configuration to IsyServiceApiSecurityAutoConfiguration.java
* enable usage of ConditionalOnClass
* adjust Tests
