+++ isy-serviceapi-core/src/test/java/de/bund/bva/isyfact/serviceapi/core/aufrufkontext/TestConfig.java
@@ -3,6 +3,7 @@ package de.bund.bva.isyfact.serviceapi.core.aufrufkontext;
 import de.bund.bva.isyfact.aufrufkontext.AufrufKontextFactory;
 import de.bund.bva.isyfact.aufrufkontext.AufrufKontextVerwalter;
 import de.bund.bva.isyfact.aufrufkontext.impl.AufrufKontextFactoryImpl;
+import de.bund.bva.isyfact.serviceapi.autoconfigure.IsyServiceApiSecurityAutoConfiguration;
 import de.bund.bva.isyfact.serviceapi.core.aufrufkontext.helper.DebugAufrufKontextVerwalter;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -10,7 +11,7 @@ import org.springframework.context.annotation.Configuration;
 import de.bund.bva.isyfact.serviceapi.core.aop.test.AufrufKontextSstTestBean;
 
 @Configuration
-@EnableAutoConfiguration
+@EnableAutoConfiguration(exclude = IsyServiceApiSecurityAutoConfiguration.class)
 public class TestConfig {
 
     @Bean
